@@ -1,7 +1,7 @@
 __title__ = 'music_manager'
 __description__ = 'Music Manager'
 __url__ = 'https://github.com/dskrypa/music_manager'
-__version__ = '2025.06.29-1'
+__version__ = '2025.07.03'
 __author__ = 'Doug Skrypa'
 __author_email__ = 'dskrypa@gmail.com'
 __copyright__ = 'Copyright 2023 Doug Skrypa'
